Renaming to VRCToggleToolkit
